+++ src/shaders.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+extern const GLchar* g_swipeShaderFragment;
+extern const GLchar* g_swipeShaderVertex;
+extern bool g_swipeShaderFailedToLoad;